Add routine to prechunk test data
